Split adaptive step and constant step base classes out of the solver module
